Moved tests inside folder. Fixed API when it's a POST with no content-type
