+++ api/knowledge/recommendations.js
@@ -82,7 +82,7 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -129,14 +129,14 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "buildId",
                 "description": "Format - int64. Unique identifier of the build",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "integer"
             }]
         };
@@ -180,7 +180,7 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -224,7 +224,7 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Operation ID",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -267,7 +267,7 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model to be fetched.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -345,14 +345,14 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "buildId",
                 "description": "Format - int64. Unique identifier of the build",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "integer"
             }]
         };
@@ -396,14 +396,14 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "buildId",
                 "description": "Format - int64. Unique identifier of the build",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "integer"
             }]
         };
@@ -447,14 +447,14 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "onlyLastRequestedBuild",
                 "description": "true to return only the last build of the model, false to return all the builds",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "boolean"
             }]
         };
@@ -502,42 +502,42 @@ Once you have created a model, you can upload your catalog data, upload your usa
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "itemIds",
                 "description": "Comma-separated list of the items to recommend for. If the active build is of type FBT, then you can send only one item. <br> Max length: 1024",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "numberOfResults",
                 "description": "Format - int32. Number of recommended items to return",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "integer"
             }, {
                 "name": "minimalScore",
                 "description": "Format - double. Minmal score, currently honored for only FBT builds.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "number"
             }, {
                 "name": "includeMetadata",
                 "description": "Future use, always false.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "boolean"
             }, {
                 "name": "buildId",
                 "description": "Format - int64. The build id to use for this recommendation request. If the number is less than 0, uses the active build of the model instead.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "integer"
             }]
         };
@@ -590,42 +590,42 @@ Note:  There is no user recommendation for an <i>FBT</i> build. If the active bu
                 "description": "Unique identifier of the model.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "userId",
                 "description": "Unique user identifier.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "numberOfResults",
                 "description": "Format - int32. Number of recommended items to return.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "integer"
             }, {
                 "name": "itemsIds",
                 "description": "The unique identifiers of the items to consider in addition to user history.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }, {
                 "name": "includeMetadata",
                 "description": "Set to false. (For future use)",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "boolean"
             }, {
                 "name": "buildId",
                 "description": "Format - int64. Build identifier, if the number is less than 0, uses the active build of the model instead.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "integer"
             }]
         };
@@ -668,7 +668,7 @@ Note:  There is no user recommendation for an <i>FBT</i> build. If the active bu
                 "description": "Operation ID",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -713,7 +713,7 @@ Note:  There is no user recommendation for an <i>FBT</i> build. If the active bu
                 "description": "Unique identifier of the model.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -759,14 +759,14 @@ Note:  There is no user recommendation for an <i>FBT</i> build. If the active bu
                 "description": "Unique identifier of the model",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "catalogDisplayName",
                 "description": "Display name of the catalog data. e.g. \"CatalogFile1\"\r\n						Only letters (A-Z, a-z), numbers (0-9), hyphens (-) and underscore (_) are allowed. Max length: 50",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -811,14 +811,14 @@ Note:  There is no user recommendation for an <i>FBT</i> build. If the active bu
                 "description": "Unique identifier of the model.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "usageDisplayName",
                 "description": "Display name of the usage data. e.g. \"UsageFile1\"<br>Only letters(A-Z, a-z), numbers(0-9), hyphens(-) and underscore(_) are allowed. <br>Max length: 50",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
