+++ api/vision/face.js
@@ -83,7 +83,7 @@ const face = ({
                     "true"
                 ],
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "boolean"
             }, {
                 "name": "returnFaceLandmarks",
@@ -94,14 +94,14 @@ const face = ({
                     "true"
                 ],
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "boolean"
             }, {
                 "name": "returnFaceAttributes",
                 "description": "Analyze and return the one or more specified face attributes in the comma-separated string like \"returnFaceAttributes=age,gender\".Supported face attributes include age, gender, headPose, smile, facialHair, and glasses. Note that each face attribute analysis has additional computational and time cost.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }]
         };
@@ -424,21 +424,21 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "userData",
                 "description": "User-specified data for any purpose. The	maximum length is 1KB.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }, {
                 "name": "targetFace",
                 "description": "A face rectangle to specify the target face to be added into the face list, in the format of \"targetFace=left,top,width,height\". E.g. \"targetFace=10,10,100,100\". No targetFace means to detect the only face in the entire image.",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }]
         };
@@ -505,7 +505,7 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.\n	",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -551,14 +551,14 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "persistedFaceId",
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -604,7 +604,7 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -649,7 +649,7 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -743,7 +743,7 @@ Face list is a group of faces, and these faces will not expire. Face list is use
                 "description": "Valid character is letter in lower case or digit or '-' or '_', maximum length is 64.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -816,28 +816,28 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "Target person that the face is added to.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "userData",
                 "description": "User-specified data for any purpose. The maximum length is 1KB. ",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }, {
                 "name": "targetFace",
                 "description": "A face rectangle to specify the target face to be added to a person, in the format of \"targetFace=left,top,width,height\". E.g. \"targetFace=10,10,100,100\". No targetFace means to detect the only face in the entire image. ",
                 "value": null,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }]
         };
@@ -900,7 +900,7 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -955,14 +955,14 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "The target personId to delete.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1009,21 +1009,21 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "Belonging person ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "persistedFaceId",
                 "description": "The face to remove. This persisted face ID is returned from <a href=\"/docs/services/563879b61984550e40cbbe8d/operations/563879b61984550f3039523b\">Person - Add a Person Face</a>.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1068,14 +1068,14 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "The target person ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1120,21 +1120,21 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "The target person ID the face belongs to.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "persistedFaceId",
                 "description": "The target face ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1177,21 +1177,21 @@ and
                 "description": "Target person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "start",
                 "description": "List persons from the least personId greater than the \"start\". It contains no more than 64 characters. Default is empty.",
                 "value": "",
                 "required": false,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "top",
                 "description": "The number of persons to list, ranging in [1, 1000]. Default is 1000.",
                 "value": 1000,
                 "required": false,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "number"
             }]
         };
@@ -1252,14 +1252,14 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "Target person's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1319,21 +1319,21 @@ and
                 "description": "The target person's belonging person group's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "personId",
                 "description": "Target person's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }, {
                 "name": "persistedFaceId",
                 "description": "Target face's ID.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1397,7 +1397,7 @@ and
                 "description": "User-provided person group ID as a string. The valid characters include numbers, english letters in lower case, '-' and '_'. The maximum length of the personGroupId is 64.\n	",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1443,7 +1443,7 @@ and
                 "description": "The ID of the person group to be deleted.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1487,7 +1487,7 @@ and
                 "description": "ID of the target person group.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1530,7 +1530,7 @@ and
                 "description": "personGroupId of target person group.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1571,14 +1571,14 @@ and
                 "description": "List person groups from the least personGroupId greater than the \"start\". It contains no more than 64 characters. Default is empty.",
                 "value": "",
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "string"
             }, {
                 "name": "top",
                 "description": "The number of person groups to list, ranging in [1, 1000]. Default is 1000.",
                 "value": 1000,
                 "required": false,
-                "kind": 2,
+                "type": "queryStringParam",
                 "typeName": "number"
             }]
         };
@@ -1625,7 +1625,7 @@ and
                 "description": "Target person group to be trained.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
@@ -1686,7 +1686,7 @@ and
                 "description": "personGroupId of the person group to be updated.",
                 "value": null,
                 "required": true,
-                "kind": 1,
+                "type": "routeParam",
                 "typeName": "string"
             }]
         };
