+++ gulpfile.js
@@ -11,7 +11,7 @@ gulp.task('pre-test', function () {
   });
   
   gulp.task('test', ['pre-test'], function () {
-    return gulp.src(['test/*.js'])
+    return gulp.src(['test/**/*.js'])
       .pipe(mocha({
           timeout: 10000
       }))
