+++ lib/api.js
@@ -107,7 +107,7 @@ function makeRequest(client, data) {
 			
 			// mandatory query params part of the path
 			(operation.parameters || []).forEach((param) => {
-				if (parameters[param.name] && param.kind == 1) {
+				if (parameters[param.name] && param.type == 'routeParam') {
 					path = path.split(`{${param.name}}`).join(parameters[param.name]);
 				}
 			});
@@ -124,7 +124,6 @@ function makeRequest(client, data) {
 			let host = operation.host || client._endpoint;
 
 			let uri = `https://${host}/${path}`;
-			console.log("URI: " + uri);
 
 			var options = {
 				uri,
@@ -136,18 +135,19 @@ function makeRequest(client, data) {
 			if (body != null) {
 				options.body = body;
 				if (headers['Content-type'] && headers['Content-type'] != 'application/json') {
-					options.json = false;
-					// POST: receive the response properly
-					options.transform = function (body, res) {
-						if (res.headers['content-type'] && res.headers['content-type'].indexOf('application/json') != -1) {
-							return JSON.parse(body);
-						}
-						if (res.headers['operation-location']) {
-							return res.headers['operation-location'];
-						}
-						return body;
-					}
+					options.json = false; //do not stringify the request body to JSON
 				}
+				// POST: receive the response properly
+				options.transform = function (body, res) {
+					if (res.headers['content-type'] && res.headers['content-type'].indexOf('application/json') != -1 && typeof(body) !== 'object') {
+						return JSON.parse(body);
+					}
+					if (res.headers['operation-location']) {
+						return res.headers['operation-location'];
+					}
+					return body;
+				}
+				
 			}
 			return rp(options);
 		})
